perf: Only backup when needed
To reduce noise when browsing backups, let's only do it when we make a
real change.  Rather than detecting when a script will be a no-op, we
are relying on the processing from the script leading to no permament
changes.  Its a bit simpler.  As of right now though, we always do
permament changes.