fix(graph): Actually delinearize commits
Before, we were making branches not stacked on branches.  This was done
for selecting transactions for rebasing.

Now, we are making it so all forks are at the end of a "stack", even if
a stack won't end in a branch.  This helps for re-linearizing.  This
should only impact transactions by making some have less useful
information.

We probably should make things delinearlized by commit by default and
then specifically choose what we want for transactions and display.