fix(cli): Dump config in gitconfig format
Unfortunately, I couldn't find a way with libgit2 to do this.  There
doesn't seem to be a way to copy values between configs to write to a
temp file.  There doesn't seem to be any in-memory writing/reading
support.

So had to do it by hand and hope its good enough.

Fixes #10