fix(cli): Change branch status precedence
Seeing a see of `ready` wasn't very informative.  I want to know what
the status is with what is pushed already, so giving that higher
precedence.

The downside is if you'd added a WIP commit, we don't muck with the
status, making it clear a `git stack --push` won't do anything.