fix(graph): Gracefully handle squash-merges
Before, when doing a `git-stack --pull`, we'd rebase the merged branch
onto master.  `git` would the automatically detect the commit is merged
and drop it.  We'd then associate the branch name with the same commit
master points to and have to delete it manually.

This tries to detect that case so we can automatically delete the stale
branch on behalf of users.

This does not help if your branch is behind master because those extra
commits will impact how we are detecting this case (tree id).

This is a spiritual revert of 80f5158e592eaeb7ead904c66a4ac2aab1daa86e,
re-implementing 892117bfc80f50d4170546829ad27920b7796bd6 to handle all
of our changes and with a little more experience and context.

This is a part of #28.