fix(cli): Change `--pull` to not perform `auto` ops
In trying to document the flags, I realized I was having a hard time
describing them.  One in particular is the role and intention of
`--pull`.  By not enabling the `auto` behaviors, I feel this makes it
clearer to describe the intention, creates a better analog to standard
git workflows, and is outright better behavior (a pull-driven rebase
doesn't get blocked on fixups, repairs, etc).