fix(cli): Do not show protect-branch merge bases
This will help reduce a lot of tree depth for people with a lot of
release branches hanging around.

This also removes the main cause of stackoverflow in #90 because we
won't be finding the common ancestor across all protected branches and
then having to recurse from that point to the leaf commits.  We still
have a problem when dealing with branches not marked as protected.

Fixes #113