feat(graph): Avoid showing large branches
Generally, people don't work on 100+ commit personal branches.  Once its
shared, we should probably consider it protected.  For now, I took a
simple way out that seems to work well for raising visibility.  Instead
of the branch being marked as protected with all of the ceremony around
that, we are marking the commits as protected.  This will skip any
current and future special processing for protected branches and color
it in a special way, helping to draw attention to it so people are more
likely to correctly configure `git-stack`.
This helps with #107.

This also gives us an implicit recursion cap.  This should be our last
item for preventing issues with stackoverflow.

Fixes #90