feat(graph): Avoid showing old branches
This is a step towards #111.  For now, we are collapsing old branches
into a single item and still showing their merge-base.  We will not try
to rebase them or process fixups.

Unlike `-count`, we are not listing the names of those branches with a
message about the age because I'm assuming we'll soon get this to the
ideal state where we list these branches like we do empty stacks.