fix(graph): Don't lose branches with --onto
We were trying to rebase the `onto` onto itself, causing us to lose
everything. but the common commits.

Overall, it makes sense that if a branch is either the `--base` or
`--onto`, that we wouldn't touch it, so let's treat them as protected.
I did this in a quick and dirty way, by making them protected branches.
Really, we should make the commits protected.