refactor(graph): Limit squash detection
Instead of looking for squashes among all protected commits, we'll only
look at the delta between the local branch and the remote.

We could check against the base but that will likely lead to a lot of
extra processing when moving between release branches, with little
likely benefit.  If real world scenarios uncover otherwise, we can
change.

This still works for `git fetch` but breaks for
`git pull` / `git stack --pull` because the local and remote commits
will be the same.