fix(graph): Update empty stacks on pull
Generally, when a developer has a branch on the same commit as `main`,
its just empty and we should also "rebase it".  Our old pull code does
not do this.  If someone does a `git fetch` followed by `git stack
--rebase`, we will now update these branches.

For when we fix all of these changes to support `git stack --pull`, this
will also take care of updating `main`.

Fixes #142