perf(git): Cache merge-bases in-memory
In the linux repo with a topic branch off of a very old commit, `git
stack` is taking around 18s.  This drops it to 12s.  We are doing a lot
of repeated merge-base calls as each level has its own version of the
question to ask.  At first, I was worried about the order of the
parameters limiting the value of the cache but we seem to be pretty
consistent in the order we provide them.

See https://github.com/gitext-rs/git-stack/discussions/216