fix(cli): Find a good-enough base when none
Originally, we would bail out if there is no base.  We might split
off branches that are very-divergent branches from their base, so we
need to make this better.  The first step was allowing it.  This is the
next where we infer a "good enough" base to probide context for the
user.