feat(run): Fail fast support
If a command failed on a commit, we won't run the command on the
descendant commits as they are likely to fail as well and someone is
likely wanting to fix the first failure.

The default can be configured by changing the alias to include
`--no-ff`.  A user can then override this with the hidden `--ff`.

Fixes #262